requirements.txt
beautifulsoup4~=4.12.3
PyMuPDF~=1.24.4
requests~=2.32.2
pytesseract~=0.3.10
pillow~=10.3.0
spire~=0.4.2
python-docx~=1.1.2